requirements.txt
aiogram==2.13
cachetools==4.2.2
python-decouple==3.4